Use a slightly more forgiving coverage config
This is what we have on st-pax and incident. By restricting it to 'CI',
it also allows a dev to explicitly allow _any_ coverage drop if they
want to as long as they commit the change to the coverage file -- if
it's an unacceptible drop, that can be caught in code review.